packages/sie_router/pyproject.toml
[project]
name = "sie-router"
version = "0.1.10"
description = "Search Inference Engine - Stateless request router for elastic cloud deployments"
requires-python = ">=3.12,<3.13"
license = { text = "Apache-2.0" }
dependencies = [
    # Web framework
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "uvloop>=0.22",

    # HTTP client for forwarding requests
    "httpx>=0.28,<1",

    # WebSocket client for worker connections
    "websockets>=14,<15",

    # NATS pub/sub for config distribution
    "nats-py>=2.9,<3",

    # Serialization (for request/response forwarding)
    "msgpack>=1.1,<2",
    "msgpack-numpy>=0.4,<1",
    "orjson>=3.10,<4",

    # Shared types (SDK is source of truth for interface types)
    "sie-sdk>=0.1",

    # Config
    "pyyaml>=6,<7",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.6,<3",

    # Observability
    "prometheus-client>=0.21,<1",

    # CLI
    "typer>=0.14,<1",
    "rich>=13,<14",

    # Hot reload (config file watching)
    "watchdog>=6,<7",
]

[project.optional-dependencies]
kubernetes = ["kubernetes>=31,<32"]
gcs = ["google-cloud-storage>=2.18,<3"]

[project.scripts]
sie-router = "sie_router.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sie_router"]

[dependency-groups]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
]
